+++ tests/models-operation-based-with-interfaces/src/test/kotlin/test/AnimalsTest.kt
@@ -5,7 +5,6 @@ import org.junit.Test
 class AnimalsTest {
   @Test
   fun animalTest() {
-    val data = animals.TestOperationQuery.Data {}
-    println(data)
+    animals.TestOperationQuery.Data {}
   }
 }
