+++ tests/sqlite/build.gradle.kts
@@ -22,6 +22,12 @@ kotlin {
         implementation(libs.apollo.normalizedcache.sqlite.incubating)
       }
     }
+    findByName("jvmTest")?.apply {
+      dependencies {
+        implementation(libs.slf4j)
+      }
+
+    }
   }
 }
